# public/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/receipt": {
      "post": {
        "operationId": "ReceiptController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReceiptDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "receipt"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "ReceiptController_root",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "receipt"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "My Nest Js API",
    "description": "The Nest JS API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateItemDto": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string"
          },
          "price": {
            "type": "number"
          }
        },
        "required": [
          "description",
          "price"
        ]
      },
      "CreateReceiptDto": {
        "type": "object",
        "properties": {
          "shopName": {
            "type": "string"
          },
          "shopAddress": {
            "type": "string"
          },
          "shopPhone": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateItemDto"
            }
          },
          "total": {
            "type": "number"
          },
          "cash": {
            "type": "number"
          },
          "change": {
            "type": "number"
          },
          "bankCardLast4Digits": {
            "type": "string"
          },
          "approvalCode": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "shopName",
          "shopAddress",
          "shopPhone",
          "items",
          "total",
          "cash",
          "change"
        ]
      }
    }
  }
}